+++ contracts/encoders/evm-encoder/src/processor/evict_msgs.rs
@@ -2,7 +2,9 @@ use alloy_sol_types::SolValue;
 use cosmwasm_std::{Binary, StdResult};
 use valence_authorization_utils::authorization::Priority;
 
-use crate::solidity_types::{EvictMsgs, ProcessorMessage, ProcessorMessageType};
+use valence_encoder_utils::processor::solidity_types::{
+    EvictMsgs, ProcessorMessage, ProcessorMessageType,
+};
 
 pub fn encode(queue_position: u64, priority: Priority) -> StdResult<Binary> {
     let message = EvictMsgs {
