+++ contracts/encoders/evm-encoder/src/processor/insert_msgs.rs
@@ -1,22 +1,23 @@
 use alloy_sol_types::SolValue;
 use cosmwasm_std::{Binary, StdError, StdResult};
 use valence_authorization_utils::authorization::{Priority, Subroutine};
-use valence_encoder_utils::msg::Message;
-
-use crate::{
-    encode_subroutine,
-    solidity_types::{ProcessorMessage, ProcessorMessageType, SendMsgs},
-    EVMLibrary,
+use valence_encoder_utils::{
+    msg::Message,
+    processor::solidity_types::{InsertMsgs, ProcessorMessage, ProcessorMessageType},
 };
+
+use crate::{encode_subroutine, EVMLibrary};
 
 pub fn encode(
     execution_id: u64,
+    queue_position: u64,
     priority: Priority,
     subroutine: Subroutine,
     messages: Vec<Message>,
 ) -> StdResult<Binary> {
-    let message = SendMsgs {
+    let message = InsertMsgs {
         executionId: execution_id,
+        queuePosition: queue_position,
         priority: priority.into(),
         subroutine: encode_subroutine(subroutine)?,
         messages: messages
@@ -29,7 +30,7 @@ pub fn encode(
     };
 
     let processor_message = ProcessorMessage {
-        messageType: ProcessorMessageType::SendMsgs,
+        messageType: ProcessorMessageType::InsertMsgs,
         message: message.abi_encode().into(),
     };
