+++ contracts/encoders/evm-encoder/src/processor/mod.rs
@@ -1,0 +1,5 @@
+pub mod evict_msgs;
+pub mod insert_msgs;
+pub mod pause;
+pub mod resume;
+pub mod send_msgs;
