+++ contracts/encoders/evm-encoder/src/processor/pause.rs
@@ -1,8 +1,7 @@
 use alloy_primitives::Bytes;
 use alloy_sol_types::SolValue;
 use cosmwasm_std::Binary;
-
-use crate::solidity_types::{ProcessorMessage, ProcessorMessageType};
+use valence_encoder_utils::processor::solidity_types::{ProcessorMessage, ProcessorMessageType};
 
 pub fn encode() -> Binary {
     let processor_message = ProcessorMessage {
