+++ packages/encoder-utils/src/lib.rs
@@ -1,1 +1,2 @@
 pub mod msg;
+pub mod processor;
