+++ packages/encoder-utils/src/processor/mod.rs
@@ -1,0 +1,1 @@
+pub mod solidity_types;
